Fix rank sort for products in collection query
