+++ saleor/graphql/product/schema.py
@@ -1,7 +1,4 @@
-from typing import List, Optional
-
 import graphene
-from graphql import GraphQLError
 
 from ...permission.enums import ProductPermissions
 from ...permission.utils import has_one_of_permissions
@@ -127,7 +124,6 @@ from .sorters import (
     CategorySortingInput,
     CollectionSortingInput,
     ProductOrder,
-    ProductOrderField,
     ProductTypeSortingInput,
     ProductVariantSortingInput,
 )
@@ -145,16 +141,7 @@ from .types import (
     ProductVariant,
     ProductVariantCountableConnection,
 )
-
-
-def search_string_in_kwargs(kwargs: dict) -> bool:
-    return bool(kwargs.get("filter", {}).get("search", "").strip()) or bool(
-        kwargs.get("search", "").strip()
-    )
-
-
-def sort_field_from_kwargs(kwargs: dict) -> Optional[List[str]]:
-    return kwargs.get("sort_by", {}).get("field") or None
+from .utils import check_for_sorting_by_rank
 
 
 class ProductQueries(graphene.ObjectType):
@@ -451,24 +438,9 @@ class ProductQueries(graphene.ObjectType):
     @staticmethod
     @traced_resolver
     def resolve_products(_root, info: ResolveInfo, *, channel=None, **kwargs):
-        if sort_field_from_kwargs(kwargs) == ProductOrderField.RANK:
-            # sort by RANK can be used only with search filter
-            if not search_string_in_kwargs(kwargs):
-                raise GraphQLError(
-                    (
-                        "Sorting by RANK is available only when using a search filter "
-                        "or search argument."
-                    )
-                )
-        if search_string_in_kwargs(kwargs) and not sort_field_from_kwargs(kwargs):
-            # default to sorting by RANK if search is used
-            # and no explicit sorting is requested
-            product_type = info.schema.get_type("ProductOrder")
-            kwargs["sort_by"] = product_type.create_container(
-                {"direction": "-", "field": ["search_rank", "id"]}
-            )
-
+        check_for_sorting_by_rank(info, kwargs)
         search = kwargs.get("search")
+
         requestor = get_user_or_app_from_context(info.context)
         has_required_permissions = has_one_of_permissions(
             requestor, ALL_PRODUCTS_PERMISSIONS
